Cleanups and simplifications.
