+++ .gitignore
@@ -4,3 +4,5 @@ bin/
 develop-eggs/
 eggs/
 *.egg-info/
+.coverage
+htmlcov/
