+++ src/zope/app/authentication/ftesting.zcml
@@ -10,13 +10,6 @@
 
   <include package="zope.securitypolicy" file="meta.zcml" />
 
-  <!-- <include package="zope.app.zcmlfiles" /> -->
-  <!-- <include package="zope.formlib" /> -->
-  <!-- <include package="zope.session" /> -->
-  <!-- <include package="zope.app.authentication" /> -->
-  <!-- <include package="zope.securitypolicy" /> -->
-  <!-- <include package="zope.login" /> -->
-
   <include package="zope.component" file="meta.zcml" />
   <include package="zope.security" file="meta.zcml" />
   <include package="zope.app.container.browser" file="meta.zcml" />
@@ -55,6 +48,12 @@
   <include package="zope.traversing" />
   <include package="zope.traversing.browser" />
 
+  <!-- Normally done in zope.app.principalannotation's bootstrap
+       subscriber -->
+  <utility
+      factory="zope.principalannotation.utility.PrincipalAnnotationUtility"
+      provides="zope.principalannotation.interfaces.IPrincipalAnnotationUtility"
+      />
 
   <!-- These normally come from zope.app.zcmlfiles/menus.zcml -->
   <browser:menu
@@ -81,6 +80,7 @@
       description="Menu of objects to be added to site management folders"
       />
 
+  <!-- This sets up session and cookie handlers in its subscribers -->
   <include package="zope.app.appsetup" />
 
   <include package="zope.app.form.browser" />
@@ -102,8 +102,7 @@
   <include package="zope.app.authentication.browser" />
 
   <!-- 401 redirect views -->
-  <!-- Not ported yet -->
-  <!-- <include package="zope.app.exception.browser" /> -->
+  <!-- normally from zope.app.exception.browser -->
   <browser:page
       for="zope.security.interfaces.IUnauthorized"
       name="index.html"
